feat: adjust permission for osp management
Refs: #994
