+++ src/administration/Administration.Service/Controllers/NetworkController.cs
@@ -22,6 +22,7 @@ using Microsoft.AspNetCore.Mvc;
 using Org.Eclipse.TractusX.Portal.Backend.Administration.Service.BusinessLogic;
 using Org.Eclipse.TractusX.Portal.Backend.Administration.Service.Models;
 using Org.Eclipse.TractusX.Portal.Backend.Framework.ErrorHandling.Web;
+using Org.Eclipse.TractusX.Portal.Backend.Framework.Models;
 using Org.Eclipse.TractusX.Portal.Backend.Framework.Web;
 using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.PortalEntities.Enums;
 using Org.Eclipse.TractusX.Portal.Backend.Web.Identity;
@@ -32,33 +33,23 @@ namespace Org.Eclipse.TractusX.Portal.Backend.Administration.Service.Controllers
 [EnvironmentRoute("MVC_ROUTING_BASEPATH", "registration/[controller]")]
 [Produces("application/json")]
 [Consumes("application/json")]
-public class NetworkController : ControllerBase
+public class NetworkController(INetworkBusinessLogic logic)
+    : ControllerBase
 {
-    private readonly INetworkBusinessLogic _logic;
-
-    /// <summary>
-    /// Creates a new instance of <see cref="NetworkController"/>
-    /// </summary>
-    /// <param name="logic">The business logic for the registration</param>
-    public NetworkController(INetworkBusinessLogic logic)
-    {
-        _logic = logic;
-    }
-
     /// <summary>
     /// Registers a partner company
     /// </summary>
     /// <param name="data">Data for the registration</param>
-    /// Example: POST: api/administration/registration/network/{externalId}/retrigger-synchronize-users
+    /// Example: POST: api/administration/registration/network/{externalId}/partnerRegistration
     /// <response code="200">Empty response on success.</response>
     [HttpPost]
     [Authorize(Policy = PolicyTypes.ValidCompany)]
-    [Authorize(Roles = "create_partner_registration")]
+    [Authorize(Roles = "configure_partner_registration")]
     [Route("partnerRegistration")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     public async Task<OkResult> PartnerRegister([FromBody] PartnerRegistrationData data)
     {
-        await _logic.HandlePartnerRegistration(data).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.HandlePartnerRegistration(data).ConfigureAwait(ConfigureAwaitOptions.None);
         return Ok();
     }
 
@@ -79,7 +70,7 @@ public class NetworkController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerSynchronizeUser([FromRoute] string externalId)
     {
-        await _logic.RetriggerProcessStep(externalId, ProcessStepTypeId.RETRIGGER_SYNCHRONIZE_USER).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.RetriggerProcessStep(externalId, ProcessStepTypeId.RETRIGGER_SYNCHRONIZE_USER).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -100,7 +91,7 @@ public class NetworkController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerCallbackOspApprove([FromRoute] string externalId)
     {
-        await _logic.RetriggerProcessStep(externalId, ProcessStepTypeId.RETRIGGER_CALLBACK_OSP_APPROVED).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.RetriggerProcessStep(externalId, ProcessStepTypeId.RETRIGGER_CALLBACK_OSP_APPROVED).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -121,7 +112,7 @@ public class NetworkController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerCallbackOspDecline([FromRoute] string externalId)
     {
-        await _logic.RetriggerProcessStep(externalId, ProcessStepTypeId.RETRIGGER_CALLBACK_OSP_DECLINED).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.RetriggerProcessStep(externalId, ProcessStepTypeId.RETRIGGER_CALLBACK_OSP_DECLINED).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -142,7 +133,7 @@ public class NetworkController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerCallbackOspSubmitted([FromRoute] string externalId)
     {
-        await _logic.RetriggerProcessStep(externalId, ProcessStepTypeId.RETRIGGER_CALLBACK_OSP_SUBMITTED).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.RetriggerProcessStep(externalId, ProcessStepTypeId.RETRIGGER_CALLBACK_OSP_SUBMITTED).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -163,7 +154,30 @@ public class NetworkController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerRemoveKeycloakUser([FromRoute] string externalId)
     {
-        await _logic.RetriggerProcessStep(externalId, ProcessStepTypeId.RETRIGGER_REMOVE_KEYCLOAK_USERS).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.RetriggerProcessStep(externalId, ProcessStepTypeId.RETRIGGER_REMOVE_KEYCLOAK_USERS).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
+
+    /// <summary>
+    /// Get OSP Company Application Detail by Company Name or Status
+    /// </summary>
+    /// <param name="page">page index start from 0</param>
+    /// <param name="size">size to get number of records</param>
+    /// <param name="companyApplicationStatusFilter">Search by company applicationstatus</param>
+    /// <param name="companyName">search by company name</param>
+    /// <param name="externalId">search by external Id</param>
+    /// <param name="dateCreatedOrderFilter">sort result by dateCreated ascending or descending</param>
+    /// <returns>OSp Company Application Details</returns>
+    /// <remarks>
+    /// Example: GET: api/administration/registration/network/companies?companyName=Car&amp;page=0&amp;size=4&amp;companyApplicationStatus=Closed <br />
+    /// Example: GET: api/administration/registration/network/companies?page=0&amp;size=4
+    /// </remarks>
+    /// <response code="200">Result as a OSP Company Application Details</response>
+    [HttpGet]
+    [Authorize(Roles = "configure_partner_registration")]
+    [Authorize(Policy = PolicyTypes.ValidCompany)]
+    [Route("companies")]
+    [ProducesResponseType(typeof(Pagination.Response<CompanyDetailsOspOnboarding>), StatusCodes.Status200OK)]
+    public Task<Pagination.Response<CompanyDetailsOspOnboarding>> GetOspCompanyDetailsAsync([FromQuery] int page, [FromQuery] int size, [FromQuery] CompanyApplicationStatusFilter? companyApplicationStatusFilter = null, [FromQuery] string? companyName = null, [FromQuery] string? externalId = null, [FromQuery] DateCreatedOrderFilter? dateCreatedOrderFilter = null) =>
+        logic.GetOspCompanyDetailsAsync(page, size, companyApplicationStatusFilter, companyName, externalId, dateCreatedOrderFilter);
 }
