+++ src/administration/Administration.Service/Controllers/RegistrationController.cs
@@ -38,19 +38,9 @@ namespace Org.Eclipse.TractusX.Portal.Backend.Administration.Service.Controllers
 [EnvironmentRoute("MVC_ROUTING_BASEPATH", "registration")]
 [Produces("application/json")]
 [Consumes("application/json")]
-public class RegistrationController : ControllerBase
+public class RegistrationController(IRegistrationBusinessLogic logic)
+    : ControllerBase
 {
-    private readonly IRegistrationBusinessLogic _logic;
-
-    /// <summary>
-    /// Creates a new instance of <see cref="RegistrationController"/>
-    /// </summary>
-    /// <param name="logic">The business logic for the registration</param>
-    public RegistrationController(IRegistrationBusinessLogic logic)
-    {
-        _logic = logic;
-    }
-
     /// <summary>
     /// Gets the company with its address
     /// </summary>
@@ -67,7 +57,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public Task<CompanyWithAddressData> GetCompanyWithAddressAsync([FromRoute] Guid applicationId) =>
-        _logic.GetCompanyWithAddressAsync(applicationId);
+        logic.GetCompanyWithAddressAsync(applicationId);
 
     /// <summary>
     /// Get Application Detail by Company Name or Status
@@ -87,7 +77,7 @@ public class RegistrationController : ControllerBase
     [Route("applications")]
     [ProducesResponseType(typeof(Pagination.Response<CompanyApplicationDetails>), StatusCodes.Status200OK)]
     public Task<Pagination.Response<CompanyApplicationDetails>> GetApplicationDetailsAsync([FromQuery] int page, [FromQuery] int size, [FromQuery] CompanyApplicationStatusFilter? companyApplicationStatusFilter = null, [FromQuery] string? companyName = null) =>
-        _logic.GetCompanyApplicationDetailsAsync(page, size, companyApplicationStatusFilter, companyName);
+        logic.GetCompanyApplicationDetailsAsync(page, size, companyApplicationStatusFilter, companyName);
 
     /// <summary>
     /// fetch all applications details with company user details.
@@ -103,7 +93,7 @@ public class RegistrationController : ControllerBase
     [Route("applicationsWithStatus")]
     [ProducesResponseType(typeof(Pagination.Response<CompanyApplicationWithCompanyUserDetails>), StatusCodes.Status200OK)]
     public Task<Pagination.Response<CompanyApplicationWithCompanyUserDetails>> GetAllCompanyApplicationsDetailsAsync([FromQuery] int page = 0, [FromQuery] int size = 15, [FromQuery] string? companyName = null) =>
-        _logic.GetAllCompanyApplicationsDetailsAsync(page, size, companyName);
+        logic.GetAllCompanyApplicationsDetailsAsync(page, size, companyName);
 
     /// <summary>
     /// Update the BPN for a Company
@@ -125,7 +115,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status409Conflict)]
     public Task UpdateCompanyBpn([FromRoute] Guid applicationId, [FromRoute] string bpn) =>
-        _logic.UpdateCompanyBpn(applicationId, bpn);
+        logic.UpdateCompanyBpn(applicationId, bpn);
 
     /// <summary>
     /// Approves the registration verification for the application with the given id
@@ -146,7 +136,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> ApproveApplication([FromRoute] Guid applicationId)
     {
-        await _logic.ApproveRegistrationVerification(applicationId).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.ApproveRegistrationVerification(applicationId).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -171,7 +161,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> DeclineApplication([FromRoute] Guid applicationId, [FromBody] RegistrationDeclineData data, CancellationToken cancellationToken)
     {
-        await _logic.DeclineRegistrationVerification(applicationId, data.Comment, cancellationToken).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.DeclineRegistrationVerification(applicationId, data.Comment, cancellationToken).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -193,7 +183,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> ProcessClearinghouseResponse([FromBody] ClearinghouseResponseData responseData, CancellationToken cancellationToken)
     {
-        await _logic.ProcessClearinghouseResponseAsync(responseData, cancellationToken).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.ProcessClearinghouseResponseAsync(responseData, cancellationToken).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -216,7 +206,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> ProcessDimResponse([FromRoute] string bpn, [FromBody] DimWalletData responseData, CancellationToken cancellationToken)
     {
-        await _logic.ProcessDimResponseAsync(bpn, responseData, cancellationToken).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.ProcessDimResponseAsync(bpn, responseData, cancellationToken).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -235,7 +225,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ChecklistDetails), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public Task<IEnumerable<ChecklistDetails>> GetChecklistForApplication([FromRoute] Guid applicationId) =>
-        _logic.GetChecklistForApplicationAsync(applicationId);
+        logic.GetChecklistForApplicationAsync(applicationId);
 
     /// <summary>
     /// Retriggers the last failed to override the clearinghouse-result
@@ -255,7 +245,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> OverrideClearinghouseChecklist([FromRoute] Guid applicationId)
     {
-        await _logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.CLEARING_HOUSE, ProcessStepTypeId.MANUAL_TRIGGER_OVERRIDE_CLEARING_HOUSE).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.CLEARING_HOUSE, ProcessStepTypeId.MANUAL_TRIGGER_OVERRIDE_CLEARING_HOUSE).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -277,7 +267,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerClearinghouseChecklist([FromRoute] Guid applicationId)
     {
-        await _logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.CLEARING_HOUSE, ProcessStepTypeId.RETRIGGER_CLEARING_HOUSE).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.CLEARING_HOUSE, ProcessStepTypeId.RETRIGGER_CLEARING_HOUSE).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -299,7 +289,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> TriggerIdentityWallet([FromRoute] Guid applicationId)
     {
-        await _logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.IDENTITY_WALLET, ProcessStepTypeId.RETRIGGER_IDENTITY_WALLET).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.IDENTITY_WALLET, ProcessStepTypeId.RETRIGGER_IDENTITY_WALLET).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -321,7 +311,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> TriggerSelfDescription([FromRoute] Guid applicationId)
     {
-        await _logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.SELF_DESCRIPTION_LP, ProcessStepTypeId.RETRIGGER_SELF_DESCRIPTION_LP).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.SELF_DESCRIPTION_LP, ProcessStepTypeId.RETRIGGER_SELF_DESCRIPTION_LP).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -344,7 +334,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> TriggerBpn([FromRoute] Guid applicationId, [FromQuery] ProcessStepTypeId processTypeId)
     {
-        await _logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.BUSINESS_PARTNER_NUMBER, processTypeId).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.BUSINESS_PARTNER_NUMBER, processTypeId).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -366,7 +356,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> ProcessClearinghouseSelfDescription([FromBody] SelfDescriptionResponseData data, CancellationToken cancellationToken)
     {
-        await _logic.ProcessClearinghouseSelfDescription(data, cancellationToken).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.ProcessClearinghouseSelfDescription(data, cancellationToken).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -385,7 +375,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(FileContentResult), StatusCodes.Status404NotFound)]
     public async Task<ActionResult> GetDocumentContentFileAsync([FromRoute] Guid documentId)
     {
-        var (fileName, content, contentType) = await _logic.GetDocumentAsync(documentId).ConfigureAwait(ConfigureAwaitOptions.None);
+        var (fileName, content, contentType) = await logic.GetDocumentAsync(documentId).ConfigureAwait(ConfigureAwaitOptions.None);
         return File(content, contentType, fileName);
     }
 
@@ -407,7 +397,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerCreateDimWallet([FromRoute] Guid applicationId)
     {
-        await _logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.IDENTITY_WALLET, ProcessStepTypeId.RETRIGGER_CREATE_DIM_WALLET).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.IDENTITY_WALLET, ProcessStepTypeId.RETRIGGER_CREATE_DIM_WALLET).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -429,7 +419,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerValidateDid([FromRoute] Guid applicationId)
     {
-        await _logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.IDENTITY_WALLET, ProcessStepTypeId.RETRIGGER_VALIDATE_DID_DOCUMENT).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.IDENTITY_WALLET, ProcessStepTypeId.RETRIGGER_VALIDATE_DID_DOCUMENT).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -451,7 +441,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerTransmitBpnDid([FromRoute] Guid applicationId)
     {
-        await _logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.IDENTITY_WALLET, ProcessStepTypeId.RETRIGGER_TRANSMIT_DID_BPN).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.IDENTITY_WALLET, ProcessStepTypeId.RETRIGGER_TRANSMIT_DID_BPN).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -473,7 +463,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerBpnCredential([FromRoute] Guid applicationId)
     {
-        await _logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.BPNL_CREDENTIAL, ProcessStepTypeId.RETRIGGER_REQUEST_BPN_CREDENTIAL).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.BPNL_CREDENTIAL, ProcessStepTypeId.RETRIGGER_REQUEST_BPN_CREDENTIAL).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -495,7 +485,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerMembershipCredential([FromRoute] Guid applicationId)
     {
-        await _logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.MEMBERSHIP_CREDENTIAL, ProcessStepTypeId.RETRIGGER_REQUEST_MEMBERSHIP_CREDENTIAL).ConfigureAwait(ConfigureAwaitOptions.None);
+        await logic.TriggerChecklistAsync(applicationId, ApplicationChecklistEntryTypeId.MEMBERSHIP_CREDENTIAL, ProcessStepTypeId.RETRIGGER_REQUEST_MEMBERSHIP_CREDENTIAL).ConfigureAwait(ConfigureAwaitOptions.None);
         return NoContent();
     }
 
@@ -517,7 +507,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> ProcessIssuerBpnResponse([FromBody] IssuerResponseData responseData, CancellationToken cancellationToken)
     {
-        await _logic.ProcessIssuerBpnResponseAsync(responseData, cancellationToken).ConfigureAwait(false);
+        await logic.ProcessIssuerBpnResponseAsync(responseData, cancellationToken).ConfigureAwait(false);
         return NoContent();
     }
 
@@ -539,7 +529,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> ProcessIssuerMembershipResponse([FromBody] IssuerResponseData responseData, CancellationToken cancellationToken)
     {
-        await _logic.ProcessIssuerMembershipResponseAsync(responseData, cancellationToken).ConfigureAwait(false);
+        await logic.ProcessIssuerMembershipResponseAsync(responseData, cancellationToken).ConfigureAwait(false);
         return NoContent();
     }
 
@@ -560,7 +550,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerDeleteIdpSharedRealm([FromRoute] Guid processId)
     {
-        await _logic.RetriggerDeleteIdpSharedRealm(processId).ConfigureAwait(false);
+        await logic.RetriggerDeleteIdpSharedRealm(processId).ConfigureAwait(false);
         return NoContent();
     }
 
@@ -581,7 +571,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerDeleteIdpSharedServiceAccount([FromRoute] Guid processId)
     {
-        await _logic.RetriggerDeleteIdpSharedServiceAccount(processId).ConfigureAwait(false);
+        await logic.RetriggerDeleteIdpSharedServiceAccount(processId).ConfigureAwait(false);
         return NoContent();
     }
 
@@ -602,7 +592,7 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerDeleteCentralIdentityProvider([FromRoute] Guid processId)
     {
-        await _logic.RetriggerDeleteCentralIdentityProvider(processId).ConfigureAwait(false);
+        await logic.RetriggerDeleteCentralIdentityProvider(processId).ConfigureAwait(false);
         return NoContent();
     }
 
@@ -623,31 +613,8 @@ public class RegistrationController : ControllerBase
     [ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status404NotFound)]
     public async Task<NoContentResult> RetriggerDeleteCentralUser([FromRoute] Guid processId)
     {
-        await _logic.RetriggerDeleteCentralUser(processId).ConfigureAwait(false);
+        await logic.RetriggerDeleteCentralUser(processId).ConfigureAwait(false);
         return NoContent();
     }
-
-    /// <summary>
-    /// Get OSP Company Application Detail by Company Name or Status
-    /// </summary>
-    /// <param name="page">page index start from 0</param>
-    /// <param name="size">size to get number of records</param>
-    /// <param name="companyApplicationStatusFilter">Search by company applicationstatus</param>
-    /// <param name="companyName">search by company name</param>
-    /// <param name="externalId">search by external Id</param>
-    /// <param name="dateCreatedOrderFilter">sort result by dateCreated ascending or descending</param>
-    /// <returns>OSp Company Application Details</returns>
-    /// <remarks>
-    /// Example: GET: api/administration/registration/network/companies?companyName=Car&amp;page=0&amp;size=4&amp;companyApplicationStatus=Closed <br />
-    /// Example: GET: api/administration/registration/network/companies?page=0&amp;size=4
-    /// </remarks>
-    /// <response code="200">Result as a OSP Company Application Details</response>
-    [HttpGet]
-    [Authorize(Roles = "view_submitted_applications")]
-    [Authorize(Policy = PolicyTypes.ValidCompany)]
-    [Route("network/companies")]
-    [ProducesResponseType(typeof(Pagination.Response<CompanyDetailsOspOnboarding>), StatusCodes.Status200OK)]
-    public Task<Pagination.Response<CompanyDetailsOspOnboarding>> GetOspCompanyDetailsAsync([FromQuery] int page, [FromQuery] int size, [FromQuery] CompanyApplicationStatusFilter? companyApplicationStatusFilter = null, [FromQuery] string? companyName = null, [FromQuery] string? externalId = null, [FromQuery] DateCreatedOrderFilter? dateCreatedOrderFilter = null) =>
-        _logic.GetOspCompanyDetailsAsync(page, size, companyApplicationStatusFilter, companyName, externalId, dateCreatedOrderFilter);
 }
