+++ src/administration/Administration.Service/appsettings.json
@@ -387,7 +387,8 @@
   },
   "Network2Network": {
     "InitialRoles": [],
-    "BasePortalAddress": ""
+    "BasePortalAddress": "",
+    "ApplicationsMaxPageSize": 20
   },
   "OnboardingServiceProvider": {
     "Username": "",
